treat TODO collections as the parent in TODOs and the dashboard
When they are a parent, any Event or Equipment is fetched as a child
object of the collection.

When they are children, the RIDs are just stored as they are already a
child of the Event or Equipment they are associated with.

For equipment and events, they are still children and will remain so.

This commit also cleans up the other models for consistent type usage.